requirements-freeze.txt
a2wsgi==1.10.7
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.20.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
asgiref==3.8.1
asyncssh==2.19.0
atpublic==5.0
attrs==24.2.0
bentoml==1.3.16
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.0
cattrs==23.1.2
celery==5.4.0
Cerberus==1.3.5
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
circus==0.18.0
click==8.1.7
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.9
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
deepmerge==2.0
Deprecated==1.2.15
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docker-pycreds==0.4.0
dpath==2.2.0
dulwich==0.22.6
dvc==3.58.0
dvc-data==3.16.7
dvc-gs==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.16.1
Flask==3.1.0
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.55.3
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.10.0
funcy==2.0
gcsfs==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grandalf==0.8
gto==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-ws==0.7.0
huggingface-hub==0.26.5
hydra-core==1.3.2
idna==3.10
importlib-metadata==6.11.0
inflection==0.5.1
iterative-telemetry==0.0.9
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
omegaconf==2.3.0
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
pip==24.0
pip-requirements-parser==32.0.1
platformdirs==4.3.6
prometheus_client==0.21.1
prompt-toolkit==3.0.36
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.3
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
questionary==2.0.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rich==12.6.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
schema==0.7.7
scikit-learn==1.6.0
scipy==1.14.1
scmrepo==3.3.9
seaborn==0.13.2
semantic-version==2.10.0
semver==3.0.2
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.6.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
simple-di==0.1.5
simple-term-menu==0.4.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sqltrie==0.11.1
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tomli_w==1.1.0
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.0
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uv==0.5.8
uvicorn==0.32.1
vine==5.1.0
voluptuous==0.15.2
wandb==0.19.1
watchfiles==1.0.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zc.lockfile==3.0.post1
zipp==3.21.0
